Move bean to two new packages
