+++ dev/com.ibm.ws.jsf.2.2_fat/test-applications/TestResourceContractsFromJar.war/src/com/ibm/ws/jsf22/fat/contractsfromjar/beans/jsf22/PickContract.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright (c)  2015  IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ */
+package com.ibm.ws.jsf22.fat.contractsfromjar.beans.jsf22;
+
+import java.io.Serializable;
+
+import javax.faces.bean.ManagedBean;
+import javax.faces.bean.SessionScoped;
+
+@ManagedBean
+@SessionScoped
+public class PickContract implements Serializable {
+
+   
+    private static final long serialVersionUID = -9057892871690903355L;
+
+    private String contract = "team";
+
+    public String getContract() {
+        return contract;
+    }
+
+    public void setContract(String contract) {
+        this.contract = contract;
+    }
+
+}
